% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.02.app_server.R
\name{app_server}
\alias{app_server}
\title{Run WASP GUI}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny app input.}

\item{output}{Shiny app output.}

\item{session}{Shiny app session information}
}
\description{
Main server function for WASP GUI.
}
