% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_VT.R
\name{calc.scaleSTDratio}
\alias{calc.scaleSTDratio}
\title{Calculate the ratio of conditional error standard deviations}
\usage{
calc.scaleSTDratio(x, zin, zout)
}
\arguments{
\item{x}{A vector of response.}

\item{zin}{A matrix containing the meaningful predictors selected from a large set of possible predictors (z).}

\item{zout}{A matrix containing the remaining possible predictors after taking out the meaningful predictors (zin).}
}
\value{
The STD ratio.
}
\description{
Calculate the ratio of conditional error standard deviations
}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
