% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{calculatePhaseDiff}
\alias{calculatePhaseDiff}
\title{Calculate phase differences from phase matrix}
\usage{
calculatePhaseDiff(phaseMatrix, refLoc, bPlot = TRUE)
}
\arguments{
\item{phaseMatrix}{A matrix with phase angles over time for considered locations.}

\item{refLoc}{A row number for reference location.}

\item{bPlot}{Logical; if \code{TRUE} plots are generated for each pair of locations.}
}
\value{
A matrix with pairwise phase differences.
}
\description{
Calculates pairwise differences between phase angles.
}
\details{
Phase differences are calculated between the reference location and all other locations. 
    The row for the reference locations will contain NAs.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
