% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{calculateWaveletSpectrum}
\alias{calculateWaveletSpectrum}
\title{Calculate wavelet spectra and phase angle matrix for time series}
\usage{
calculateWaveletSpectrum(bwn, locs = 1:nrow(bwn), units = 12,
  periodStep = 1/100, bRemoveNulls = FALSE)
}
\arguments{
\item{bwn}{A matrix of values for which the wavelet spectrum is desired.}

\item{locs}{A vector of rows (i.e. locations) to be used in the analysis.}

\item{units}{Number of time units in one period (e.g. months in a year) (X axis)}

\item{periodStep}{Resolution of time component (i.e. periods) (Y axis)}

\item{bRemoveNulls}{Logical; if \code{TRUE} removes all NULL objects from the final list.}
}
\value{
Returns a list of wavelet outputs.
}
\description{
Computes wavelet spectra and creates a matrix of phase angles.
}
\details{
Rows with uniform zero incidence will be ignored and excluded from the resulting list.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
