% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modwt_vt.R
\name{class.modwt}
\alias{class.modwt}
\title{Convert matrix to modwt class object}
\usage{
class.modwt(B.modwt, J, wf, boundary)
}
\arguments{
\item{B.modwt}{A matrix of modwt decomposition.}

\item{J}{Specifies the depth of the decomposition. This must be a number less than or equal to log(length(x),2).}

\item{wf}{Name of the wavelet filter to use in the decomposition.}

\item{boundary}{Character string specifying the boundary condition. If boundary=="periodic" the default, then the vector you decompose is assumed to be periodic on its defined interval, if boundary=="reflection", the vector beyond its boundaries is assumed to be a symmetric reflection of itself.}
}
\value{
A class of "modwt" object
}
\description{
Convert matrix to modwt class object
}
