% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{col_grad}
\alias{col_grad}
\title{Gradient Color Palette}
\usage{
col_grad(scm = 1)
}
\arguments{
\item{scm}{Gradient scheme to use
(1 = viridis, 2 = yellow/brown, 3 = blue/red-a, 4 = blue/red-b)}
}
\value{
Vector of colors to replace default gradient color scale.
}
\description{
Returns 12 colors from the viridis color palette.
}
\examples{

col_grad()

}
