% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{col_univ}
\alias{col_univ}
\title{Universal Color Palette}
\usage{
col_univ()
}
\value{
Vector of colors to replace default discrete color scale.
}
\description{
Combines the npg, aaas, and lancet ggsci palettes for use with datasets
containing up to 36 groups.
}
\examples{

 # col_univ()

}
