% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{correctPhase}
\alias{correctPhase}
\title{Correct phase angles}
\usage{
correctPhase(vector)
}
\arguments{
\item{vector}{A vector of phase angles.}
}
\value{
A vector with phase angles that fall within [-pi, pi] range.
}
\description{
Function to correct for spikes while computing phase difference.
}
\details{
While perfrming arithmetical operations with phase angles, one could end up with values outside [-pi, pi] range.
    This function adds or subtracts additional \code{2 * pi} for correction.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
