% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.params.R
\name{create_proc_params}
\alias{create_proc_params}
\title{Define Processing Parameters}
\usage{
create_proc_params(d_path, study_md)
}
\arguments{
\item{d_path}{Character string indicating the path
to a set of CellRanger files for data processing.}

\item{study_md}{A data frame containing sample
metadata variables specific to an individual study.}
}
\value{
Data frame containing a list of parameters
to use for scRNA-Seq data processing by Seurat.
}
\description{
Creates a data frame of processing parameters
used in scRNA-Seq data processing by Seurat.
}
\examples{

# # Dataset input parameters
# list_params <- create_proc_params(
#   "data/",
#   data.frame(
#     # ID column name
#     # (splits by underscore, should be listed
#     # first in folder name as in 's01_1_KO_a')
#     "Code" = unlist(
#        lapply(
#          strsplit(
#            basename(list.files("data/")),
#            "_",
#            fixed = T
#          ),
#          "[",
#          1
#        )
#      ),
#     # 1st metadata column (include in CellRanger folder name)
#     "Knockout" = as.factor(
#                    ifelse(
#                      grepl("NG",basename(list.files("data/"))),
#                      "ctrl",
#                      "KO"
#                    )
#                  ),
#     # 2nd metadata column
#     "Region" = as.factor(
#                  ifelse(
#                    grepl("LAE",basename(list.files("data/"))),
#                    "1",
#                    "2"
#                  )
#                ),
#     # 3rd metadata column (add/remove columns as needed)
#     "Time" = as.factor(
#                ifelse(
#                  grepl("D28",basename(list.files("data/"))),
#                  "a",
#                  "b"
#                )
#              )
#   )
# )

}
