% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{createPhaseMatrix}
\alias{createPhaseMatrix}
\title{Create a matrix of phase angles}
\usage{
createPhaseMatrix(wts, period = 1, significantPowerThreshold = 0)
}
\arguments{
\item{wts}{List of wavelet transform outputs.}

\item{period}{The period of interest.}

\item{significantPowerThreshold}{A threshold for power, e.g. if equals to 5, phase angles with corresponding power < 5 will be NAs.}
}
\value{
Returns a matrix with phase angles.
}
\description{
Creates a matrix of phase angles for a specified period
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
