% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{createPowerLines}
\alias{createPowerLines}
\title{Create power lines}
\usage{
createPowerLines(wts)
}
\arguments{
\item{wts}{List of wavelet transform outputs.}
}
\value{
Returns a matrix where each row is a power line.
}
\description{
Creates power lines
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
