% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{createSlidingPhaseMatrix}
\alias{createSlidingPhaseMatrix}
\title{Calculate average phase angles}
\usage{
createSlidingPhaseMatrix(phaseMatrix, window = 12, bSort = TRUE)
}
\arguments{
\item{phaseMatrix}{A matrix of phase angles.}

\item{window}{The size of the moving average window.}

\item{bSort}{Logical; if \code{TRUE} sorts the output matrix according to the average phase (over time).}
}
\value{
Returns a matrix with average phase angle differences from other locations.
}
\description{
Creates a matrix with mean phase differences between regions for a moving time period.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
