% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data.AWAP.2.5}
\alias{data.AWAP.2.5}
\title{Sample data: AWAP rainfall data over Australia}
\usage{
data(data.AWAP.2.5)
}
\description{
A dataset containing 1320 rows (data length) and 252 columns (grids).
}
\keyword{datasets}
