% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data.CI}
\alias{data.CI}
\title{Sample data: Climate indices strongly influencing Australia climate}
\usage{
data(data.CI)
}
\description{
A dataset containing 1332 rows (data length) and 6 columns (indices).
}
\keyword{datasets}
