% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data.HL}
\alias{data.HL}
\title{Sample data: Hysteresis loop}
\usage{
data(data.HL)
}
\description{
A dataset containing 3 lists: a vector of response (x), a matrix of 9 potential predictors (dp) with each column containing one potential predictor, and a vector of true predictor numbers.
}
\keyword{datasets}
