% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data.Rossler}
\alias{data.Rossler}
\title{Sample data: Rossler system}
\usage{
data(data.Rossler)
}
\description{
A list with four vectors (n=100,000) named time, x, y and z containing the time, the x-components, the y-components and the z-components of the Rossler system, respectively.
}
\keyword{datasets}
