% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data.SW1}
\alias{data.SW1}
\title{Sample data: Sinewave model 1 (SW1):\deqn{ x_{t}= sin(2pi*f*t)+eps}}
\usage{
data(data.SW1)
}
\description{
A dataset containing 3 lists: a vector of response (x), a matrix of 9 potential predictors (dp) with each column containing one potential predictor, and a vector of true predictor numbers.
}
\keyword{datasets}
