% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColorsFromNumbers}
\alias{getColorsFromNumbers}
\title{Get colors for numeric vector values}
\usage{
getColorsFromNumbers(x, colors = c("blue", "white", "red"),
  range = range(x))
}
\arguments{
\item{x}{A numeric vector of values.}

\item{colors}{A vector of colors for the palette.}

\item{range}{A range of values for color picking.}
}
\value{
Returns a vector with colors.
}
\description{
Function to convert numbers into colors.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
