% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{getPeaksMatrix}
\alias{getPeaksMatrix}
\title{Get matrix of peaks}
\usage{
getPeaksMatrix(cases, start, period = 12)
}
\arguments{
\item{cases}{A matrix of case series.}

\item{start}{The start timepoint of each season.}

\item{period}{The period of seasons (e.g. 12 for monthly data, and 52 for weekly).}
}
\value{
Returns a matrix with peak timing for each season and location.
}
\description{
Creates a matrix with peak timing for each season.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
