% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{Ind_AWAP.2.5}
\alias{Ind_AWAP.2.5}
\title{Sample data: Index of AWAP grids with no missing data}
\usage{
data(Ind_AWAP.2.5)
}
\description{
A dataset containing 145 numbers.
}
\keyword{datasets}
