% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knnregl1cv}
\alias{knnregl1cv}
\title{Leave one out cross validation.}
\usage{
knnregl1cv(x, z, k = 0, pw)
}
\arguments{
\item{x}{A vector of response.}

\item{z}{A matrix of predictors.}

\item{k}{The number of nearest neighbours used. The default is 0, indicating Lall and Sharma default is used.}

\item{pw}{A vector of partial weights of the same length of z.}
}
\value{
A vector of L1CV estimates of the response.
}
\description{
Leave one out cross validation.
}
\references{
Lall, U., Sharma, A., 1996. A Nearest Neighbor Bootstrap For Resampling Hydrologic Time Series. Water Resources Research, 32(3): 679-693.

Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
