% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{lat_lon.2.5}
\alias{lat_lon.2.5}
\title{Sample data: Latitude and longitude of AWAP grids}
\usage{
data(lat_lon.2.5)
}
\description{
A dataset containing 252 rows (grids) and 2 columns (lat and lon).
}
\keyword{datasets}
