% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.03.mod_import.R
\name{mod_data_import_server}
\alias{mod_data_import_server}
\title{Data Import Module Server}
\usage{
mod_data_import_server(id)
}
\arguments{
\item{id}{A string. The module namespace id.}
}
\value{
A reactive containing the imported Seurat object.
}
\description{
Data Import Module Server
}
