% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.03.mod_import.R
\name{mod_data_import_ui}
\alias{mod_data_import_ui}
\title{Data Import Module UI}
\usage{
mod_data_import_ui(id)
}
\arguments{
\item{id}{A string. The module namespace id.}
}
\description{
Data Import Module UI
}
