% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.04.mod_filter.R
\name{mod_filter_server}
\alias{mod_filter_server}
\title{Filter Module Server}
\usage{
mod_filter_server(id, data, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{A string. The module namespace id.}

\item{data}{A reactive Seurat object from mod_data_import_server.}
}
\value{
A reactive filtered Seurat object.
}
\description{
Filter Module Server
}
