% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.05.mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{Plot Module Server}
\usage{
mod_plot_server(id, data, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{A string. The module namespace id.}

\item{data}{A reactive Seurat object from mod_filter_server.}
}
\description{
Plot Module Server
}
