% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.06.mod_table.R
\name{mod_summary_table_server}
\alias{mod_summary_table_server}
\title{Summary Table Module Server}
\usage{
mod_summary_table_server(id, data)
}
\arguments{
\item{id}{A string. The module namespace id.}

\item{data}{A reactive Seurat object from mod_filter_server.}
}
\description{
Summary Table Module Server
}
