% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.06.mod_table.R
\name{mod_summary_table_ui}
\alias{mod_summary_table_ui}
\title{Summary Table Module UI}
\usage{
mod_summary_table_ui(id)
}
\arguments{
\item{id}{A string. The module namespace id.}
}
\description{
Summary Table Module UI
}
