% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_variance.R
\name{non.bdy}
\alias{non.bdy}
\title{Replace Boundary Wavelet Coefficients with Missing Values (NA).}
\usage{
non.bdy(x, wf, method = c("dwt", "modwt", "mra"))
}
\arguments{
\item{x}{DWT/MODWT/AT object}

\item{wf}{Character string; name of wavelet filter}

\item{method}{Either dwt or modwt or mra}
}
\value{
Same object as x only with some missing values (NA).
}
\description{
Replace Boundary Wavelet Coefficients with Missing Values (NA).
}
\examples{

}
\references{
Cornish, C. R., Bretherton, C. S., & Percival, D. B. (2006). Maximal overlap wavelet statistical analysis with application to atmospheric turbulence. Boundary-Layer Meteorology, 119(2), 339-374.
}
