% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{obs.mon}
\alias{obs.mon}
\title{Sample data: NCEP reanalysis data averaged over Sydney region}
\usage{
data(obs.mon)
}
\description{
A dataset containing 720 rows (data length) and 7 columns (atmospheric variables).
}
\keyword{datasets}
