% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_VT.R
\name{pic.calc}
\alias{pic.calc}
\title{Calculate PIC}
\usage{
pic.calc(X, Y, Z, mode, wf)
}
\arguments{
\item{X}{A vector of response.}

\item{Y}{A matrix of new predictors.}

\item{Z}{A matrix of pre-existing predictors that could be NULL if no prior predictors exist.}

\item{mode}{A mode of variance transfomration, i.e., MRA, MODWT, or AT}

\item{wf}{Wavelet family}
}
\value{
A list of 2 elements: the partial mutual information (pmi), and partial informational correlation (pic).
}
\description{
Calculate PIC
}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.

Galelli S., Humphrey G.B., Maier H.R., Castelletti A., Dandy G.C. and Gibbs M.S. (2014) An evaluation framework for input variable selection algorithms for environmental data-driven models, Environmental Modelling and Software, 62, 33-51, DOI: 10.1016/j.envsoft.2014.08.015.
}
