% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{plotAnnualCurves}
\alias{plotAnnualCurves}
\title{Plot annual curves}
\usage{
plotAnnualCurves(cases, start, period = 12, perc = 0.95)
}
\arguments{
\item{cases}{A matrix of case series.}

\item{start}{The start timepoint of each season.}

\item{period}{The period of seasons (e.g. 12 for monthly data, and 52 for weekly).}

\item{perc}{CI range for the mean peak timing.}
}
\description{
Plots seasonal curves for each location with peaks and mean peak timing.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
