% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{plotAveragePhaseLag}
\alias{plotAveragePhaseLag}
\title{Plot mean phase lags averaged over time}
\usage{
plotAveragePhaseLag(slideMat, perc = 0.95)
}
\arguments{
\item{slideMat}{A matrix of phase angles.}

\item{perc}{Percentage envelope for CI around the mean.}
}
\description{
Plot mean phase lags averaged over time.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
