% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{plotMeanPeakTiming}
\alias{plotMeanPeakTiming}
\title{Plot mean peak timing}
\usage{
plotMeanPeakTiming(pMat, start, period = 12, perc = 0.95,
  bBootstrap = FALSE)
}
\arguments{
\item{pMat}{A matrix with peaks for each location and season.}

\item{start}{The start timepoint of each season.}

\item{period}{The period of seasons (e.g. 12 for monthly data, and 52 for weekly).}

\item{perc}{CI range for the mean peak timing.}

\item{Logical;}{if \code{TRUE} bootstrap will be used.}
}
\value{
Returns a named vector of mean peaks for each location.
}
\description{
Plot mean peak timing for all locations.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
