% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{plotMeanPhaseLagOverTime}
\alias{plotMeanPhaseLagOverTime}
\title{Plot mean phase lags from other locations}
\usage{
plotMeanPhaseLagOverTime(slideMat, sel = 1:nrow(slideMat), colors = c("red",
  "yellow", "green", "blue"), legendSize = 1, legendNcol = 1, cex = 0,
  xlim = c(1, ncol(slideMat)), ylim = c(-pi, pi), xlab = "Time",
  ylab = "Phase lag from other regions", ...)
}
\arguments{
\item{slideMat}{A matrix of phase angles.}

\item{sel}{Vector of locations to use for the plot.}

\item{colors}{Vector of colors to create a palette for different locations.}

\item{legendSize}{Size of the legend.}

\item{legendNcol}{Number of columns in the legend.}

\item{cex}{.}

\item{xlim}{.}

\item{ylim}{.}

\item{xlab}{.}

\item{ylab}{.}

\item{\dots}{Additional graphical parameters.}
}
\description{
Plot mean phase lags from other locations.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
