% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{plotMedianPhaseLag}
\alias{plotMedianPhaseLag}
\title{Plot median phase lags averaged over time}
\usage{
plotMedianPhaseLag(slideMat, perc = 0.95)
}
\arguments{
\item{slideMat}{A matrix of phase angles.}

\item{perc}{Percentage envelope for quantiles around the median.}
}
\description{
Plot median phase lags averaged over time.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
