% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{plotPowerLines}
\alias{plotPowerLines}
\title{Plot power lines}
\usage{
plotPowerLines(powerLines)
}
\arguments{
\item{powerLines}{Matrix with wavelet power vectors for each location.}
}
\description{
Plots power lines
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
