% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{plotWaveletClusterDendrogram}
\alias{plotWaveletClusterDendrogram}
\title{Plot wavelet cluster dendrogram}
\usage{
plotWaveletClusterDendrogram(w.arr.dis, k = 4,
  labels = rownames(w.arr.dis$dist.mat),
  title = "Wavelet cluster dendrogram")
}
\arguments{
\item{w.arr.dis}{A matrix with distances between wavelet spectra.}

\item{k}{A number of clusters of locations.}

\item{labels}{A vector of labels for the tree tips (i.e. locations).}

\item{title}{A title for the generated graph.}
}
\value{
An object of class \code{hclust} which describes the resulting tree.
}
\description{
Generates and plots an hierarchical tree for locations based on pairwise distances between their
    wavelet spectra.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
