% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{plotWaveletGraph}
\alias{plotWaveletGraph}
\title{Plot wavelet spectrum summary graph}
\usage{
plotWaveletGraph(wts, cases, i, units = 12, period = 1,
  periodRange = c(0.2, 4), bLog = FALSE)
}
\arguments{
\item{wts}{A list of wavelet spectrum objects.}

\item{cases}{A matrix with case series.}

\item{i}{A row number (i.e. location) for plotting.}

\item{units}{Number of time units in one period (e.g. months in a year) (X axis)}

\item{period}{The period we are interested in.}

\item{periodRange}{The range of periods to plot.}

\item{bLog}{Logical; if \code{TRUE} adds a line of logged cases.}
}
\description{
Plots a summary graph with three subgraphs: timeline of case series, wavelet spectrum and 
    overall wavelet power.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
