% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{plotWaveletSpectrum}
\alias{plotWaveletSpectrum}
\title{Plot wavelet spectrum}
\usage{
plotWaveletSpectrum(wtfoc, units = 12, period = 1, periodRange = c(0.2,
  4), title = "")
}
\arguments{
\item{wtfoc}{Wavelet spectrum object.}

\item{units}{Number of time units in one period (e.g. months in a year) (X axis)}

\item{period}{The period we are interested in.}

\item{periodRange}{The range of periods to plot (Y axis).}

\item{title}{Main title for the plot.}
}
\description{
Plots wavelet spectrum
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
