% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_VT.R
\name{pw.calc}
\alias{pw.calc}
\title{Calculate Partial Weight}
\usage{
pw.calc(x, z, cpyPIC)
}
\arguments{
\item{x}{A vector of response.}

\item{z}{A matrix containing identified predictors of x.}

\item{cpyPIC}{Partial informational correlation (cpyPIC).}
}
\value{
A vector of partial weights(pw) of the same length of z.
}
\description{
Calculate Partial Weight
}
\examples{

}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
