% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_VT.R
\name{r2.boot}
\alias{r2.boot}
\title{R2 threshold by re-sampling approach}
\usage{
r2.boot(z.vt, x, prob)
}
\arguments{
\item{z.vt}{Identified independent variables}

\item{x}{Response or dependent variable}

\item{prob}{Probability with values in [0,1].}
}
\value{
A quantile assosciated with prob.
}
\description{
R2 threshold by re-sampling approach
}
