% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{rain.mon}
\alias{rain.mon}
\title{Sample data: Rainfall station data over Sydney region}
\usage{
data(rain.mon)
}
\description{
A dataset containing 732 rows (data length) and 15 columns (stations).
}
\keyword{datasets}
