% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.00.app_launcher.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{runApp}}, such as
\code{port} or \code{launch.browser}}
}
\description{
Launch the Shiny application
}
