% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.0.wasp_gui.R
\name{run_wasp_gui}
\alias{run_wasp_gui}
\title{Run WASP GUI}
\usage{
run_wasp_gui(...)
}
\arguments{
\item{...}{Arguments passed to the WASP GUI.}
}
\description{
Launches the WASP GUI. Useful for interactive analysis
of several commonly used WASP functions.
}
