% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_barplot}
\alias{sc_barplot}
\title{Cell Proportion Bar Plot}
\usage{
sc_barplot(
  dat,
  ct_col = "CellType",
  xvar = "Group",
  yvar = "Proportion",
  show_counts = TRUE,
  count_var = "Total.Cells"
)
}
\arguments{
\item{dat}{An input data frame containing counts.}

\item{ct_col}{Cell Type column or metadata variable to plot.}

\item{xvar}{An additional metadata variable for stratifying
the input data.}

\item{yvar}{Cell count variable name.}
}
\value{
A bar plot grouped by a specific metadata column.
}
\description{
Generates bar plot summarizing the number of cells per each
type for all treatment groups.
}
\examples{

# sc_barplot(dat = d)

}
