% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.5.cellchat.R
\name{sc_cc_chrd}
\alias{sc_cc_chrd}
\title{CellChat Chord Diagram}
\usage{
sc_cc_chrd(
  ccdf,
  title1,
  title2,
  cc_type,
  col_g = NULL,
  sel_pw = NULL,
  sel_src = NULL,
  sel_tar = NULL,
  spl_pw = FALSE,
  pw = 18,
  ph = 18,
  fs1 = 0.6,
  fd1 = 1.8,
  lgx = 0.1,
  lgy = 0.25,
  cc_dir = "analyze/"
)
}
\arguments{
\item{ccdf}{A CellChat data frame generat.}

\item{title1}{Plot title, provided as a character string.}

\item{cc_type}{Type of plot to use (either "total" of "comp"). Selecting
"total" plots the total number of interactions between cell types, whereas
"comp" compares the fold difference between two treatment groups.}

\item{sel_src}{vector for filtering source communication cell types.}

\item{sel_tar}{vector for filtering target communication cell types.}

\item{spl_pw}{Should the data be filtered by signaling pathway? If TRUE,
pw_sel must be specified.}

\item{pw}{Plot width (in cm).}

\item{ph}{Plot height (in cm).}

\item{fs1}{Font size (between 0.1 and 1).}

\item{fd1}{Gap between cell type labels and plot (generally between 1 and 3).}

\item{lgx}{x coordinate for legend position.}

\item{lgy}{y coordinate for legend position.}

\item{cc_dir}{Directory for saving output plot.}

\item{pw_sel}{Character string of a pathway to use for filtering the
provided CellChat object.}

\item{g_name}{Group name to split data by if cc_type is "comp."}
}
\value{
A chord diagram displaying either pathway or individual
ligand-receptor interactions.
}
\description{
Creates a chord diagram from a provided CellChat results data frame.
}
\examples{

# sc_cc_chrd(
#   ccdf = ccm[[1]],
#   title1 = "plot.cc.LR.sec",
#   title2 = "L-R Interactions: Secretory",
#   cc_type = "total",
#   sel_tar = c("5.Secretory", "16.Secretory"),
#   pw = 12,
#   ph = 12,
#   fs1 = 0.6,
#   fd1 = 2.8
# )

}
