% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.5.cellchat.R
\name{sc_cc_run}
\alias{sc_cc_run}
\title{CellChat Based Cell-cell Communication}
\usage{
sc_cc_run(so, asy1 = "SCT", g_name, s_name, ct_col = "CellType", parl = TRUE)
}
\arguments{
\item{so}{A Seurat object. Must contain a gene expression assay.}

\item{asy1}{RNA assay to use.}

\item{g_name}{Group name for comparing cell-cell communication between
treatments, provided as a character string. If only one group is present,
include an additional variable in the selected Seurat data set indicating
that all cells belong to the same group.}

\item{s_name}{Variable name containing individual sample names.}

\item{ct_col}{Cell type column name to use in the selected Seurat object.}

\item{parl}{Should CellChat be run in parallel?}
}
\value{
A data frame including either pathway or individual
ligand-receptor interactions.
}
\description{
Conducts CellChat for a selected Seurat object, either provided as
a whole data set or split by a treatment variable for comparison.
}
\examples{

# ccmerge <- sc_cc_run(
#   so = d,
#   asy1 = "RNA",
#   g_name = "Group",
#   s_name = "Code",
#   ct_col = "CellType"
# )

}
