% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.calculate.chromvar.R
\name{sc_chromvar}
\alias{sc_chromvar}
\title{ChromVAR}
\usage{
sc_chromvar(
  out_file,
  so,
  file_jaspar,
  asy = "ufy.peaks",
  genome_use = "hg38",
  bsgenome_sel = BSgenome.Hsapiens.UCSC.hg38
)
}
\arguments{
\item{out_file}{Output RDS file name containing Seurat object.}

\item{so}{Input Seurat object. Must have a unified peaks assay to
run properly.}

\item{file_jaspar}{Path to a JASPAR PFM .txt file downloaded from
https://jaspar.elixir.no/downloads/.}

\item{asy}{Name of unified ATAC peak assay in Seurat object.}

\item{genome_use}{Name of genome to use for creating motif matix.}

\item{bsgenome_sel}{BSgenome library name (provided as object).}
}
\value{
A Seurat object containing a chromVAR assay.
}
\description{
Calculates and adds a chromVAR assay to the selected Seurat object.
}
\examples{

# d1 <- sc_chromvar(
#   out_file = "data_replication_cohort.rds",
#   so = d1,
#   file_jaspar = "ref/JASPAR2026_CORE_vertebrates.txt"
# )

}
