% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.4.coverage.plot.r
\name{sc_coverage_plot2}
\alias{sc_coverage_plot2}
\title{scATAC-Seq Coverage Plot 2}
\usage{
sc_coverage_plot2(so, dref, asy1, g_name, bp_window, md_list1)
}
\arguments{
\item{so}{An object of class Seurat. Must contain an ATAC assay.}

\item{dref}{Path to a .gtf file containing reference gene annotations.}

\item{asy1}{ATAC peaks assay to use.}

\item{g_name}{Gene to plot, provided as a character string.}

\item{bp_window}{Numeric value indicating the number of base pairs to
extend the plotting window on each end of the selected gene's location.
Useful for visualizing peaks corresponding to neighboring genes.}

\item{md_list1}{Character vector of up to 2 metadata variables for
stratifying peak data.}
}
\value{
A coverage plot including the specified gene track and all other
genes present within the specified window.
}
\description{
Generates a coverage plot from a Signac ChromatinAssay.
Requires scATAC-Seq peak information and a reference GRanges object
for plotting gene positions.
}
\examples{

# p_cov <- sc_coverage_plot(
#   readRDS("analysis/data.annotated.withTFs.rds"),
#   rtracklayer::import(
#    "ref/gencode.v45.primary_assembly.annotation.gtf"
#   ),
#   "TMEM45A",
#   20000,
#   c("CellType", "Airway")
# )

}
