% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.DGEA.R
\name{sc_diff}
\alias{sc_diff}
\title{DGEA and DA}
\usage{
sc_diff(
  so,
  asy = "SCT",
  slot1 = "counts",
  ct = "CellType",
  var_id = "Code",
  mtd = "nebula",
  form1,
  cores = 1
)
}
\arguments{
\item{so}{A Seurat object}

\item{asy}{Character string providing the name of the assay
to use for differential analysis.}

\item{slot1}{Character string selecting the slot from the
Seurat object to pull from the chosen assay.}

\item{ct}{CellType column name.}

\item{var_id}{Sample ID column name.}

\item{mtd}{DGEA/DA method to use (either "MAST", "nebula",
"lme", "limma", or "wilcox"). Differential accessbility tests
run on chromVAR assays default to "wilcox".
Set to nebula by default.}

\item{form1}{Formula to use for MAST generalized linear model. Usually
consists of two terms, the first of which is the treatment group column
name and the second of which is the column indicating the number of features
present per cell.}

\item{cores}{Number of available cores to use if running
in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A list of DGEA results per cell type for the chosen group comparison,
including genes missing fold changes and cell type DGEA results
with errors.
}
\description{
Performs differential gene expression/accessibility analysis
per cell type from a Seurat Object.
}
\examples{

# scdgea <- sc_diff(
#   so = d,
#   form1 = c("Group", "nFeature_SCT")
# )

}
