% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.enrichment.R
\name{sc_enrich_plot}
\alias{sc_enrich_plot}
\title{Enrichment Plot}
\usage{
sc_enrich_plot(drich, plist = NULL, top_pw = 10)
}
\arguments{
\item{drich}{A sc_enrichment output result.}

\item{plist}{(optional) Provide a vector of pathway names for
plotting enrichment results.}

\item{top_pw}{Number of pathways to plot, ranked by p-value.}
}
\value{
An enrichment plot displaying the effect size and p-values
from a GSEA object.
}
\description{
Plots GSEA results generated by sc_enrichment. By default, the
top n pathways are plotted unless otherwise specified.
}
\examples{

# sc_enrich_plot(drich = enr_res)

}
