% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.enrichment.R
\name{sc_enrichment}
\alias{sc_enrichment}
\title{Gene Set Enrichment Analysis}
\usage{
sc_enrichment(
  mtd = "fgsea",
  pw_source = "Reactome",
  gmt = NULL,
  pw_cstm = NULL,
  colg = "Gene",
  colid = "SetID",
  colct = "CellType",
  colfc = "log2FC",
  deg,
  core_num = 0
)
}
\arguments{
\item{mtd}{Enrichment test method (set to "fgsea" by default).}

\item{pw_source}{Pathway reference source
(either "Reactome", "GO", or "Custom").
Please note that if the source is "Custom,"
a custom gene set list file must be specified by pw_cstm.}

\item{gmt}{Gene matrix transposed file containing either Reactome or
GO gene set IDs.}

\item{pw_cstm}{A text file consisting of at least two columns,
including gene names and the custom gene set IDs. Must be specified
if pw_source is "Custom," but is otherwise optional. The gene sets
are automatically appended to the specified source if pw_source is
"Reactome" or "GO."}

\item{colg}{Gene name column in custom gene list.}

\item{colid}{Gene set ID column in custom gene list.}

\item{colct}{Cell type column in DGEA results table, if applicable.}

\item{colfc}{Fold change column in DGEA results table.}

\item{deg}{A table of DGEA results generated by sc_diff or a
table of DGEA results containing p-values, fold changes, and gene names.
If performing GSEA for single-cell datasets, the DGEA results should
also contain an additional column specifying the cell type each test
was performed in.}

\item{core_num}{Number of cores to use for parallel processing
(default is 0).}
}
\value{
A data frame containing GSEA results based on the
specified DGEA results table.
}
\description{
Performs gene set enrichment analysis based on DGEA results.
}
\examples{

# sc_enrichment(
#   gmt = "ReactomePathways.gmt",
#   pw_cstm = "gref_v01.txt",
#   deg = "dgea_results.txt"
# )

}
