% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.params.R
\name{sc_format_gtf}
\alias{sc_format_gtf}
\title{Format GENCODE Gene Annotation}
\usage{
sc_format_gtf(path_ref = "ref/", path_gtf = "gencode.v50.annotation.gtf")
}
\arguments{
\item{path_ref}{Reference files path.}

\item{path_gtf}{Name of gtf file within reference directory.}
}
\value{
A formatted gene annotation object.
}
\description{
Creates a data frame of processing parameters
used in multiome data processing by Seurat.
}
\examples{

# sc_format_gtf()

}
