% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_heatmap}
\alias{sc_heatmap}
\title{General Heatmap}
\usage{
sc_heatmap(
  so = NULL,
  deg_plot = NULL,
  deg_sig = FALSE,
  filt_deg = TRUE,
  top_deg = 10,
  l_cstm = NULL,
  asy = "SCT",
  cl_var = "CellType",
  var_cl = NULL,
  h_w = 40,
  h_h = 20,
  fs_c = 6,
  fs_r = 8,
  cl_c = FALSE,
  cl_r = FALSE,
  rot_c = 45,
  rc = TRUE,
  rn = TRUE,
  col1 = col_grad(scm = 5),
  mark_dir = "analyze/",
  mark_tab = NULL,
  transp = FALSE,
  var_as_factor = FALSE,
  ret_input = FALSE,
  hmlinewidth = 0
)
}
\arguments{
\item{so}{A Seurat object.}

\item{deg_plot}{Generate heatmap based on DGEA/DA results instead
of expression/activity scores from a Seurat object. Supports DGEA/DA
result tables generated from sc_diff.}

\item{deg_sig}{Should only genes with statistically significant
differential expression/activity be plotted?}

\item{filt_deg}{If TRUE, filters genes from a Seurat object based
on differential expression in at least one group/cluster.}

\item{top_deg}{Top DEGs to plot if filtering data by DEGs.}

\item{l_cstm}{(optional) Filter data based on custom gene list.}

\item{asy}{Assay to use (ex. "RNA").}

\item{cl_var}{Character string containing the name
of the cluster variable for cell type predictions.}

\item{var_cl}{Custom gene list variable for heatmap annotation.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{rc}{Show column names?}

\item{rn}{Show row names?}

\item{col1}{Gradient color scheme to use.}

\item{mark_dir}{Directory for saving the calculated marker gene table.}

\item{mark_tab}{Path to an existing marker table for skipping the
calculation of marker genes if l_cstm is not provided.}

\item{transp}{Transpose heatmap input matrix (FALSE by default).}

\item{var_as_factor}{Should cluster variable be converted to factor?}

\item{ret_input}{Returns input heatmap matrix instead of the heatmap plot.}

\item{hmlinewidth}{Numeric value for heatmap cell border width.}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a Seurat Object,
which can be filtered based on differentially
expressed genes.
}
\examples{

# p_hmap <- sc_top10_marker_heatmap(
#   so = d_recluster[["data"]],
#   cl_var = "recluster"
# )

}
