% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_heatmap_deg}
\alias{sc_heatmap_deg}
\title{Top-10 DEG Heatmap}
\usage{
sc_heatmap_deg(
  l_deg,
  so,
  asy = "RNA",
  cl_var = "CellType",
  hm_w = 30,
  hm_h = 15,
  fs_c = 6,
  fs_r = 8,
  c_name
)
}
\arguments{
\item{l_deg}{A list of DGEA results returned by sc.DGEA().}

\item{so}{An object of class Seurat.}

\item{asy}{Character string providing the name of the assay to use.}

\item{cl_var}{Character string containing the name
of the clustering variable.}

\item{hm_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{hm_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{c_name}{Comparison name for plot title, provided as a character string.}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a Seurat Object and marker gene
list based on the top-10 marker genes for each cluster.
}
\examples{

# p.heatmap <- sc_top10_de_da_heatmap(
#   dgea_output,
#   d_annotated,
#   "RNA",
#   "seurat.clusters",
#   18,
#   24,
#   6,
#   8,
#   "grp2 vs. grp1"
# )

}
