% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.integrate.files.R
\name{sc_integrate_data}
\alias{sc_integrate_data}
\title{Batch Integration of scRNA-Seq and Multiome Datasets}
\usage{
sc_integrate_data(
  dir_proc = "processing/",
  l_so,
  l_par,
  proc_mode,
  parl = FALSE,
  core_num = NULL,
  res_clus = 0.5
)
}
\arguments{
\item{dir_proc}{Working directory for processing samples.}

\item{l_so}{List of processed Seurat objects to be integrated
from any sc_process function.}

\item{l_par}{List of processing parameters obtained from a
sc_params function.}

\item{proc_mode}{Processing mode to be used
(either "scRNA-Seq" or "multiome").}

\item{parl}{Logical indicating whether processing should be run
in parallel. Set to FALSE if running sequentially.}

\item{core_num}{Number of available cores
to use if running in parallel.}

\item{res_clus}{Seurat clustering resolution.}
}
\value{
A Seurat object containing integrated data for all samples
present in a scRNA-Seq or multiome experiment.
}
\description{
Integrates a list of samples processed as Seurat objects.
}
\examples{

# d_integrated <- sc_integrate_data(
#   l_so = scdata,
#   l_par = scparam,
#   proc_mode = "multiome"
# )

}
