% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.integrate.files.R
\name{sc_integration_qc}
\alias{sc_integration_qc}
\title{scRNA-Seq Integration Quality Control}
\usage{
sc_integration_qc(so, cl_var)
}
\arguments{
\item{so}{Integrated Seurat object.}

\item{cl_var}{Clustering variable provided as a
character string (generally use "seurat_clusters").}
}
\value{
A panel of violin plots providing QC
measures for an integrated Seurat object.
}
\description{
Displays the feature number, average counts, and percentage of reads
derived from the mitochondrial genome given an integrated Seurat object.
}
\examples{

# d.integrated <- sc_integrate_data(d.integrated)

}
