% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.params.R
\name{sc_multiome_params}
\alias{sc_multiome_params}
\title{Define Processing Parameters}
\usage{
sc_multiome_params(study_md, gtf_path, fa_path)
}
\arguments{
\item{study_md}{A data frame containing sample
metadata variables specific to an individual study.}

\item{gtf_path}{Character string providing the path
to a GENCODE gene annotation file.}

\item{fa_path}{Character string providing the path
to a GENCODE FASTA genome file.}
}
\value{
A list of parameters to use for processing
10X multiome files.
}
\description{
Creates a data frame of processing parameters
used in scRNA-Seq data processing by Seurat.
}
\examples{

# # Dataset input parameters
# l_params <- sc_multiome_params(
#   data.frame(
#     # ID column name
#     "Code" = basename(list.files("data/")),
#     # 1st metadata column (include in CellRanger folder name)
#     "Group" = as.factor(
#       ifelse(
#         grepl("DD", basename(list.files("data/"))),
#         "NonCF",
#         "CF"
#       )
#     ),
#   ),
#   "ref/gencode.v45.primary_assembly.annotation.gtf",
#   "ref/GRCh38.primary_assembly.genome.fa"
# )

}
