% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.2.top.level.fun.r
\name{sc_panel_gex}
\alias{sc_panel_gex}
\title{UMAP panels}
\usage{
sc_panel_gex(
  so,
  slot1 = "umap",
  var_ct = "CellType",
  var_md,
  pos_leg = c(0.85, 0.25)
)
}
\arguments{
\item{so}{A Seurat object.}

\item{slot1}{Name of the dimension reduction slot to plot.}

\item{var_ct}{Cell type variable name.}

\item{var_md}{Selected gene to plot.}

\item{pos_leg}{Legend position provided as a vector (default is "none").}
}
\value{
A series of UMAPs and violin plot with specified overlays.
}
\description{
Generates a panel of UMAPs from a Seurat object.
}
\examples{

# p_umap <- sc_panel_gex(so = d, var_md = "SFTPB")

}
