% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.pca.R
\name{sc_pca}
\alias{sc_pca}
\title{Seurat PCA}
\usage{
sc_pca(so, slot1)
}
\arguments{
\item{so}{An object on class Seurat.}

\item{slot1}{Character string indicating which assay to use for PCA.}
}
\value{
A Seurat object containing PCA results.
}
\description{
Performs PCA on a Seurat object.
}
\examples{

# d_integrated <- sc_pca(d_integrated, "RNA")

}
