% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.pca.R
\name{sc_pca_plot}
\alias{sc_pca_plot}
\title{scRNA-Seq PCA Loadings Plot}
\usage{
sc_pca_plot(so, md_list, red1)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md_list}{A vector of character strings indicating
metadata columns for overlaying on a loadings plot.}

\item{red1}{Reduction to use for plotting PCA components.}
}
\value{
A series of loadings plots with specified metadata overlays.
}
\description{
Generates a panel of loadings plots given
a Seurat object containing PCA results.
}
\examples{

# p_pca <- sc_pca_plot(d_integrated,c("col1","col2","col3"), "pca_cor")

}
