% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.enrichment.R
\name{sc_plot_enrichment}
\alias{sc_plot_enrichment}
\title{Enrichment Plots}
\usage{
sc_plot_enrichment(l_enr, en_type)
}
\arguments{
\item{l_enr}{A data frame containing enrichment results generated by sc.enrichment().}

\item{en_type}{Are the provided enrichment results based on the Gene Ontology database or a custom gene set? Type "GO"
for Gene Ontology Results and "cstm" for custom gene set enrichment results.}
}
\value{
An enrichment plot depicting the most significant gene sets for a given comparison and cell type.
}
\description{
Plots the top 10 terms from a provided set of enrichment results.
}
\examples{

# # Custom gene set example
# sc_plot_enrichment(
#   enr[["Results"]][enr[["Results"]][["CellType"]] == "2.Secretory",],
#   "cstm"
# )
# # Gene ontology example
# sc_plot_enrichment(
#   enr[enr[["Description"]] == "GO_BP_2.Secretory",],
#   "GO"
# )
# 

}
