% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc_plot_rho}
\alias{sc_plot_rho}
\title{Plot Sample Contamination Fraction}
\usage{
sc_plot_rho(df_par)
}
\arguments{
\item{df_par}{Data frame of updated parameters used for data processing.}
}
\value{
A scatter plot indicating the individual
and average contamination fractions for all samples.
}
\description{
Generates a scatter plot indicating the individual
and average contamination fractions for all samples.
}
\examples{

# sc_plot_rho(list_params)

}
