% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.celltype.predictions.R
\name{sc_predict}
\alias{sc_predict}
\title{Cell Type Prediction (HLCA/Azimuth)}
\usage{
sc_predict(
  so,
  ref1 = "lungref",
  f_size = 10000,
  cl_var = "seurat_clusters",
  md_list,
  parl = TRUE,
  core_num = 2
)
}
\arguments{
\item{so}{A Seurat object.}

\item{ref1}{Azimuth reference annotation ("lungref" by default).}

\item{f_size}{Future size (in MB), provided as a numeric value.
For larger datasets, set to 5000 or higher.}

\item{cl_var}{Name of the cluster variable used for cell type predictions.}

\item{md_list}{A vector selecting metadata columns
stratifying the prediction score output table.}

\item{parl}{Should predictions be run in parallel? (TRUE/FALSE)}

\item{core_num}{Number of cores to use if parl is TRUE.}
}
\value{
A list containing a Seurat object with predicted clusters and
QC tables/plots to evaluate prediction performance.
}
\description{
Cell type predictions of a Seurat Object using the built-in
Azimuth Human Lung Cell Atlas reference annotation.
}
\examples{

# pred1 <- sc_predict(
#   so = d,
#   md_list = c("Code")
# )

}
