% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.celltype.predictions.R
\name{sc_predict2}
\alias{sc_predict2}
\title{Cell Type Prediction (Custom)}
\usage{
sc_predict2(
  so,
  ref1,
  f_size = 10000,
  cl_var = "seurat_clusters",
  ct_col = "CellType",
  asy1 = "RNA",
  md_list,
  parl = TRUE,
  core_perc = 0.25
)
}
\arguments{
\item{so}{A Seurat object.}

\item{ref1}{Reference Seurat object containing annotated cell types.}

\item{f_size}{Future size (in MB), provided as a numeric value.}

\item{cl_var}{Cluster variable for cell type predictions.}

\item{ct_col}{Column to use from the reference dataset for annotating
the query dataset.}

\item{asy1}{Assay to use for comparing the reference and query dataset.}

\item{md_list}{Vector of metadata variables for stratifying proportion table.}

\item{parl}{Should predictions be run in parallel? (TRUE/FALSE)}

\item{core_perc}{Percentage of cores to use if parl is TRUE.}
}
\value{
A list containing a Seurat object with predicted clusters and
QC tables/plots to evaluate prediction performance.
}
\description{
Predicts cell type identities of clusters in a Seurat object
from a previously annotated Seurat object.
}
\examples{

# pred1 <- sc_predict2(
#   so = d,
#   ref1 = d_ref,
#   md_list = c("Code"),
# )

}
