% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc_process_batch}
\alias{sc_process_batch}
\title{Batch Processing of CellRanger Files}
\usage{
sc_process_batch(df_par, rho_adj, m_cell, m_feat, parl, core_perc)
}
\arguments{
\item{df_par}{Data frame of parameters to use for data processing.}

\item{rho_adj}{Numeric value indicating a proportion to scale
rho values calculated during the ambient RNA removal step. 10% (0.1)
is generally suitable for most data sets.}

\item{m_cell}{Threshold for including features in a Seurat Object.
Run ?Seurat::CreateSeuratObject() for more details.}

\item{m_feat}{Threshold for including cells in a Seurat Object.
Run ?Seurat::CreateSeuratObject() for more details.}

\item{parl}{Logical indicating whether processing should be run in
parallel (Linux and WSL2 only). Set to FALSE if running sequentially.}

\item{core_perc}{Percentage of available cores to use if running
in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A processed list of sample files converted into Seurat
objects with a summary list of QC and processing details.
}
\description{
Processes a list of scRNA-Seq samples for data integration.
}
\examples{

# list_data <- sc_process_batch(
# # parameter list
# list_params,
# # adj.rho proportion
# 0.1,
# # minimum cells per feature
# 5,
# # minimum features per cell
# 200
# )

}
