% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc_process_file}
\alias{sc_process_file}
\title{Process CellRanger Data Files}
\usage{
sc_process_file(df_par, i, rho_adj = 0.1, m_cell = 5, m_feat = 200)
}
\arguments{
\item{df_par}{Data frame of parameters to use for data processing.}

\item{i}{Numeric value for sample ID.}

\item{rho_adj}{Numeric value indicating a proportion to scale
rho values calculated during the ambient RNA removal step. 10% (0.1)
is generally suitable for most data sets.}

\item{m_cell}{Threshold for including features in a Seurat Object.
Run ?Seurat::CreateSeuratObject() for more details.}

\item{m_feat}{Threshold for including cells in a Seurat Object.
Run ?Seurat::CreateSeuratObject() for more details.}
}
\value{
A processed sample file converted into a Seurat object
with a summary list of QC and processing details.
}
\description{
Processes a single sample into a Seurat object for data integration.
}
\examples{

# proc_data <- sc_process_file(
#   df_par = list_params,
#   i = 1
# )

}
