% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.2.recluster.R
\name{sc_recluster}
\alias{sc_recluster}
\title{Reclustering Analysis}
\usage{
sc_recluster(
  title1,
  so,
  rc_type,
  ct_col,
  ct,
  slot1 = "RNA",
  md_list,
  prop_list,
  dim1 = 50,
  batch_cor = TRUE,
  cor_col,
  res1 = 0.5,
  h_w = 38,
  h_h = 18,
  fs_c = 6,
  fs_r = 9,
  cl_c = FALSE,
  cl_r = FALSE,
  rot_c = 45,
  col1 = col_grad(scm = 3)
)
}
\arguments{
\item{title1}{Name of cell type(s) being reclustered. This information
is used for filenames generated by reclustering analysis.}

\item{so}{An object of class Seurat.}

\item{rc_type}{Data set type being reclustered (either "GEX" or "Mult").}

\item{ct_col}{Column containing cell type information.}

\item{ct}{Character strings separated by vertical bars "|"
indicating cell types to factor in reclustering.
The provided Seurat object is then subsetted to include
only clusters from the provided group names.}

\item{slot1}{Assay from Seurat object subset to use (ex. "RNA").}

\item{md_list}{A vector of character strings indicating metadata
columns for overlaying on a loadings plot.}

\item{prop_list}{Subset of md_list to use for stratifying cell type
proportion columns.}

\item{dim1}{Number of dimensions to use in PCA.}

\item{batch_cor}{Should Harmony based batch correction be performed?
This parameter is automatically set to TRUE for multiome datasets.}

\item{cor_col}{Variables to account for in batch correction.}

\item{res1}{Clustering resolution to use for final cluster determination.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{col1}{Gradient color scheme to use
(must be exactly 4 colors in length).}
}
\value{
A reclustered Seurat Object with summary plots, proportion tables,
and marker gene lists.
}
\description{
Performs reclustering and basic visualization
of a subsetted Seurat object subset.
Subsets are determined based on cell groups.
}
\examples{

# d_recluster <- sc_recluster(
#   title1 = "basal",
#   so = d_gene,
#   rc_type = "GEX",
#   ct_col = "CellType",
#   ct = "Bas|Cyc|Sup",
#   slot1 = "RNA",
#   md_list = c("Code", "Airway", "Batch"),
#   prop_list = c("Code", "Airway"),
#   dim1 = 30,
#   cor_col = c("Batch", "Code"),
#   col1 = col_hmap
# )

}
