% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.celltype.predictions.R
\name{sc_save_pred}
\alias{sc_save_pred}
\title{Save Predicted Data}
\usage{
sc_save_pred(so_pred, file1, dir1)
}
\arguments{
\item{so_pred}{A list generated by either sc_predict() or sc_predict2().}

\item{file1}{File name for saving predicted results list}

\item{dir1}{Directory for saving files
(relative to the current working directory).}
}
\value{
Individual RDS and prediction QC files.
}
\description{
Saves a predicted list object in the chosen directory.
}
\examples{

# sc_save_pred(
#   so_pred = pred1,
#   file1 = "test",
#   dir1 = "analysis/"
# )

}
