% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.4.coverage.plot.r
\name{sc_seqanno_plot}
\alias{sc_seqanno_plot}
\title{scATAC-Seq Gene Track}
\usage{
sc_seqanno_plot(so, dref, g_name1, bp_window)
}
\arguments{
\item{so}{An object of class Seurat. Must contain an ATAC assay.}

\item{dref}{Path to a .gtf file containing reference gene annotations.}

\item{g_name1}{Gene to plot, provided as a character string.}

\item{bp_window}{Numeric value indicating the number of base pairs to
extend the plotting window on each end of the selected gene's location.
Useful for visualizing peaks corresponding to neighboring genes.}
}
\value{
A sequence plot including the specified gene track and all other
genes present within the specified window.
}
\description{
Generates a gene track from a Signac ChromatinAssay.
Requires scATAC-Seq peak information and a reference GRanges object
for plotting gene positions.
}
\examples{

# p_cov <- sc_seqanno_plot(
#   d,
#   rtracklayer::import(
#    "ref/gencode.v45.primary_assembly.annotation.gtf"
#   ),
#   "TMEM45A",
#   20000
# )

}
