% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_top10_marker_heatmap}
\alias{sc_top10_marker_heatmap}
\title{Top-10 Marker Gene Heatmap}
\usage{
sc_top10_marker_heatmap(
  so,
  title1,
  asy = "sct",
  cl_var = "CellType",
  h_w = 32,
  h_h = 20,
  fs_c = 6,
  fs_r = 8,
  cl_c = FALSE,
  cl_r = FALSE,
  rot_c = 45,
  col1 = col_grad(scm = 3)
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{asy}{Assay to use (ex. "RNA").}

\item{cl_var}{Character string containing the name
of the cluster variable for cell type predictions.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{col1}{Gradient color scheme to use
(must be exactly 4 colors in length).}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a Seurat Object and
marker gene list based on the top-10 marker genes for each cluster.
}
\examples{

# p_hmap <- sc_top10_marker_heatmap(
#   so = d_recluster[["data"]],
#   cl_var = "recluster"
# )

}
