% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_top10_marker_heatmap_rc}
\alias{sc_top10_marker_heatmap_rc}
\title{Top-10 Marker Gene Heatmap (Reclustered)}
\usage{
sc_top10_marker_heatmap_rc(
  title1,
  sorc,
  asy,
  slot1,
  cl_var,
  h_w,
  h_h,
  fs_c,
  fs_r,
  cl_c,
  cl_r,
  rot_c,
  col1,
  sc_d = TRUE,
  nmark = 10
)
}
\arguments{
\item{title1}{Celltype name to include in output files.}

\item{sorc}{An object of class Seurat.}

\item{asy}{Assay type (either "GEX" or "Mult").}

\item{slot1}{Assay to use from Seurat object subset.}

\item{cl_var}{Character string containing the name
of the cluster variable for cell type predictions.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{col1}{Gradient color scheme to use
(must be exactly 4 colors in length).}

\item{sc_d}{Scale input data?}

\item{nmark}{Number of top markers to use.}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a reclustered Seurat Object and
marker gene list based on the top-10 marker genes for each cluster.
}
\examples{

# p_umap <- sc_top10_marker_heatmap_rc(
#   "secretory",
#   d_annotated,
#   "GEX",
#   "sct",
#   "seurat_clusters",
#   18,
#   24,
#   6,
#   8,
#   TRUE,
#   TRUE,
#   col_grad()[c(3, 6, 9, 12)]
# )

}
