% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_top_motif_heatmap}
\alias{sc_top_motif_heatmap}
\title{Top Differentially Active Motif Heatmap}
\usage{
sc_top_motif_heatmap(
  so,
  mot_m,
  mot_list = NULL,
  filt_ct = FALSE,
  ct_name = NULL,
  cl_var = NULL,
  split_var = FALSE,
  inc_var = FALSE,
  list_var = NULL,
  top_n = 10,
  ord1 = FALSE,
  ord_c = NULL,
  h_w = 44,
  h_h = 14,
  fs_c = 10,
  fs_r = 10,
  cl_c = TRUE,
  cl_r = TRUE,
  col1 = col_grad
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{mot_m}{Differential activity results object.}

\item{mot_list}{(optional) A vector of motif names to plot on the
heatmap. Note that supplying a list will override filtering based
on the top n motifs for a specified activity results object.}

\item{filt_ct}{Logical indicating whether the data should
be filtered based on cell type}

\item{ct_name}{Character string containing the cell type name to filter.}

\item{cl_var}{Character string containing the name
of the clustering variable.}

\item{split_var}{Logical indicating whether the cluster variable
should be stratified by additional group variables.}

\item{inc_var}{Logical indicating whether two separate clustering variables
should be plotted in a single figure (useful for comparing summarized and
split expression together). Uses list_var if TRUE.}

\item{list_var}{(Optional) A vector of character strings
indicating the name(s) of up to two group variables
for stratifying plot points.}

\item{top_n}{Number of motifs to use (filters top-100 motifs per assay or
specified cell type and subsequently filters by scaled activity score).}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{col1}{Gradient color scheme to use.}
}
\value{
A ComplexHeatmap object containing a top motif heatmap.
}
\description{
Generates a heatmap from a Seurat Object and
differential activity analysis result.
}
\examples{

# tf_heatmap <- sc_top_motif_heatmap(
#   # Seurat object
#   so = d,
#   # Differential activity results
#   mot_m = diff_output_activity,
#   # Filter based on CellType?
#   filt_ct = TRUE,
#   # Cell type name
#   ct_name = "11.Secretory",
#   # Clustering column
#   cl_var = "CellType",
#   # Split by additional variable(s)?
#   split_var = TRUE,
#   # Additional variable(s) to split plot (if split_var = TRUE)
#   list_var = c("Airway"),
#   # Number of motifs to use
#   top_n = 50,
#   # Heatmap width
#   h_w = 36,
#   # Heatmap height
#   h_h = 12,
#   # Column font size
#   fs_c = 6,
#   # Row font size
#   fs_r = 8,
#   # Gradient color scheme
#   col_grad()[c(3, 6, 9, 12)]
# )

}
