% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap_panel}
\alias{sc_umap_panel}
\title{scRNA-Seq UMAP Plot}
\usage{
sc_umap_panel(so, md_list, slot1, leg_x = 0.9, leg_y = 0.25)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md_list}{A vector of character strings indicating
metadata columns for overlaying on a loadings plot.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}
}
\value{
A series of UMAPs with specified metadata overlays.
}
\description{
Generates a panel of UMAPs given a
Seurat object containing PCA and UMAP results.
}
\examples{

# p_umap <- sc_umap_panel(d_integrated,c("col1","col2","col3"),"wnn.umap")

}
