% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap_panel_act}
\alias{sc_umap_panel_act}
\title{scATAC-Seq Activity UMAP}
\usage{
sc_umap_panel_act(
  so,
  md_var,
  da_df,
  tf_name,
  col_scheme,
  col_names,
  leg_x,
  leg_y,
  slot1
)
}
\arguments{
\item{so}{A Seurat object.}

\item{md_var}{A character string indicating the clustering
column for overlaying on a UMAP plot.}

\item{da_df}{Differential activity results data frame.}

\item{tf_name}{Transcription factor name, provided as a character string.}

\item{col_scheme}{Colors to use for overlaying metadata variables.}

\item{col_names}{Names for assigning colors from col_scheme.}

\item{leg_x}{Legend x-axis position.}

\item{leg_y}{Legend y-axis position.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}
}
\value{
A UMAP plot with points indicating the activity of a specific
transcription factor motif.
}
\description{
Generates a series of plots given a Seurat scATAC-Seq assay.
}
\examples{

# p_umap_act <- sc_umap_panel_act(
#   d,
#   "CellType",
#   diff.activity.output,
#   "GHRL1",
#   col_univ()[1:length(levels(d@meta.data[[md_var]]))],
#   c(levels(d@meta.data[[md_var]]))
#   0.95,
#   0.95,
#   "wnn.umap.cor"
# )

}
