% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap_panel_gene}
\alias{sc_umap_panel_gene}
\title{Visualize Individual Gene Expression}
\usage{
sc_umap_panel_gene(
  so,
  asy1 = "sct",
  md_var,
  g_name,
  col_scheme,
  col_names,
  leg_x = 0.9,
  leg_y = 0.2,
  slot1 = "umap",
  col1 = col_grad(),
  plot_comb = FALSE,
  out1 = "umap_gex"
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{asy1}{Assay to use; either "sct" (default) or "chromvar".}

\item{md_var}{A character string indicating
the clustering column for overlaying on a UMAP plot.}

\item{g_name}{A character string indicating the gene name.}

\item{col_scheme}{The color scheme to be used for
distinguishing between groups, provided as a vector.}

\item{col_names}{A vector of the same length as the
provided color scheme for assigning colors to each group.}

\item{leg_x}{A numeric value indicating the placement
of the figure legend on the x-axis.}

\item{leg_y}{A numeric value indicating the placement
of the figure legend on the y-axis.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}

\item{col1}{Color scheme to use for expression gradient.}

\item{plot_comb}{Should all panels be displayed? (TRUE/FALSE)}

\item{out1}{Panel to return if plot_comb is FALSE
(either "umap_gex" or "vio_gex").}
}
\value{
A series of plots stored as a ggplot2 object
for visualizing cluster gene expression.
}
\description{
Generates a series of plots to visualize
individual gene expression per cluster.
}
\examples{

# p_umap <- sc_umap_panel_gene(
#  so = d,
#  md_var = "CellType",
#  g_name = "CFTR",
#  col_scheme = col_univ(),
#  col_names = c("group1","group2"),
#  slot1 = "umap"
# )

}
