% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap_panel_gene_list}
\alias{sc_umap_panel_gene_list}
\title{Visualize Gene List Expression}
\usage{
sc_umap_panel_gene_list(
  list_g,
  so,
  md_var,
  col_scheme,
  col_names,
  leg_x,
  leg_y,
  parl,
  core_perc,
  slot1
)
}
\arguments{
\item{list_g}{A vector containing a list of genes
to be plotted for a given Seurat object.}

\item{so}{An object of class Seurat.}

\item{md_var}{A character string indicating the clustering
column for overlaying on a UMAP plot.}

\item{col_scheme}{The color scheme to be used for distinguishing
between groups, provided as a vector.}

\item{col_names}{A vector of the same length as the provided
color scheme for assigning colors to each group.}

\item{leg_x}{A numeric value indicating the placement
of the figure legend on the x-axis.}

\item{leg_y}{A numeric value indicating the placement
of the figure legend on the y-axis.}

\item{parl}{Logical indicating whether processing should
be run in parallel (Linux and WSL2 only).
Set to FALSE if running sequentially.}

\item{core_perc}{Percentage of available cores to use if
running in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}
}
\value{
A list of plots saved as ggplot2 objects for
visualizing cluster gene expression.
}
\description{
Generates a series of plots to visualize
the expression of a list of genes per cluster.
}
\examples{

# sc_umap_panel_gene_list(
#  list_genes,
#  d_seurat,
#  "seurat_clusters",
#  col_vec,
#  col_vec_names,
#  0.95,
#  0.95,
#  TRUE,
#  0.5,
#  "umap"
# )

}
