% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap_standard}
\alias{sc_umap_standard}
\title{Standard UMAP Plot}
\usage{
sc_umap_standard(
  so,
  md_var,
  slot1 = NULL,
  dims1 = "2D",
  col1 = col_univ(),
  pos_leg = "none"
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md_var}{A character string indicating the clustering
column for overlaying on a UMAP plot.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}

\item{dims1}{Should the data be plotted in 2 or 3 dimensions?}

\item{col1}{Color scheme to use for plot overlay.}

\item{pos_leg}{Legend position (specify x and y coordinates or type "none")
(either "2D" or "3D").}
}
\value{
A UMAP plot with points grouped by a specific metadata column.
}
\description{
Generates a single UMAP plot with a specific
metadata overlay for a Seurat object.
}
\examples{

# p_umap <- sc_umap_standard(
#   so = d_integrated,
#   md_var = "CellType",
#   slot1 = "wnn.umap"
# )

}
