% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_vio_panel_gene}
\alias{sc_vio_panel_gene}
\title{Visualize Individual Gene Expression}
\usage{
sc_vio_panel_gene(
  so,
  md_vars,
  g_name,
  col_scheme = col_univ(),
  col_names = NULL,
  leg_x = 0.95,
  leg_y = 0.95,
  asy1 = "RNA"
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md_vars}{A character string indicating
the clustering column for overlaying on a UMAP plot.}

\item{g_name}{A character string indicating the gene name.}

\item{col_scheme}{The color scheme to be used for
distinguishing between groups, provided as a vector.}

\item{col_names}{A vector of the same length as the
provided color scheme for assigning colors to each group.}

\item{leg_x}{A numeric value indicating the placement
of the figure legend on the x-axis.}

\item{leg_y}{A numeric value indicating the placement
of the figure legend on the y-axis.}

\item{asy1}{Default assay for plotting violin plot ("RNA" by default).
for visualizing cluster gene expression.}
}
\description{
Generates a series of plots to visualize
individual gene expression per cluster.
}
\examples{

# p_umap <- sc_vio_panel_gene(
#   so = d, md_vars = c("CellType"), g_name = "SFTPC"
# )

}
