% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_violin}
\alias{sc_violin}
\title{Violin Plot}
\usage{
sc_violin(so, ct_col = "CellType", gene_col, pos_leg = "none")
}
\arguments{
\item{so}{An input Seurat object.}

\item{ct_col}{Cell Type column or metadata variable to plot.}

\item{gene_col}{A vector of one or more genes to plot.}

\item{pos_leg}{Legend position (specify x and y coordinates or type "none").}
}
\value{
A violin plot grouped by a specific metadata column.
}
\description{
Generates violin plot(s) with a specific
metadata overlay for a Seurat object
containing gene expression/chromatin accessibility data.
}
\examples{

# p_umap <- sc_violin(
#   so = d,
#   gene_col = c("DAPI", "EpCAM")
# )

}
