% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.2.top.level.fun.r
\name{sc_visualize}
\alias{sc_visualize}
\title{Top-Level Visualization Function}
\usage{
sc_visualize(ptype, p_params)
}
\arguments{
\item{ptype}{Plot type, selected from one of the following:
"umap_gex_list", "umap_atac_list", "umap_std", "vol_std" ...}

\item{p_params}{List of plot-specific parameters supplied to the function.
Note that function defaults are used when input parameters are missing.}
}
\value{
Plot(s) corresponding to the chosen plot type.
}
\description{
General utility function used to toggle between plot types. All necessary
R objects (datasets, gene lists, etc.) must be in the global
environment to work properly and are defined
}
\examples{

# p1 <- sc_visualize(
#   ptype = "umap_gex_list"
# )

}
