% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.volplot.R
\name{sc_volcano}
\alias{sc_volcano}
\title{Volcano Plot}
\usage{
sc_volcano(
  l_deg,
  ct = NULL,
  comp_name = NULL,
  comp_filt = FALSE,
  gene_name = "GENE",
  filt_lab = NULL,
  diff_col = "log2FC",
  cl_col = NULL,
  alt_col = NULL,
  p_col = "H.qval",
  p_cut = 0.05,
  f_cut = 0.25,
  f_lim = 5,
  y_limit = 50,
  x_title = NULL,
  pt_size = 2,
  lab_size = 4,
  mx_overlap = 18,
  ret_input = FALSE
)
}
\arguments{
\item{l_deg}{A data frame of DGEA results in the same format as
outputs generated by sc_diff.}

\item{ct}{A character string pattern for matching to specific cell types.}

\item{comp_name}{Comparison name, provided as a character string.}

\item{comp_filt}{Filter data by comparison name.}

\item{gene_name}{Column name containing gene or motif names.}

\item{filt_lab}{(optional) Select labels to include on plot.}

\item{diff_col}{Column name containing effect size.}

\item{cl_col}{Optional variable name for coloring significant
points based on class.}

\item{alt_col}{Optional named vector of colors for manually
specifying colors is cl_col is provided.}

\item{p_col}{Column name containing adjust p-values.}

\item{p_cut}{Numeric value for p-value cutoff to indicate significance.}

\item{f_cut}{Fold change cutoff indicating high effect size.}

\item{f_lim}{Numeric value for fold change limits on the x-axis.}

\item{y_limit}{Upper-bound -log10 p-value to display on the y-axis.}

\item{x_title}{X-axis legend title, provided as a character string.}

\item{pt_size}{Numeric value for point size.}

\item{lab_size}{Numeric value for label size.}

\item{mx_overlap}{Numeric value passed to EnhancedVolcano for
calculating the maximum number of overlapping labels to display.}

\item{ret_input}{Return plot input instead of plot.}
}
\value{
A volcano plot for the chosen cell type and treatment comparison.
}
\description{
Generates a volcano plot from a formatted DGEA results data frame for
a chosen comparison and/or cell type.
}
\examples{

# p_vol <- sc_volcano(
#   l_deg = diff.output.activity,
#   ct = "1.Secretory",
#   comp_name = "KO vs. Control",
#   gene_name = "gene",
#   filt_lab = grepl("NKX|FOXA", diff.output.activity[["gene"]]),
#   diff_col = "avg_diff",
#   p_col = "p_val_adj",
#   p_cut = 0.005,
#   f_cut = 0.25,
#   f_lim = 6,
#   y_limit = 100,
#   x_title = "x-axis title"
# )

}
