% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{spatialCorrelationFunction}
\alias{spatialCorrelationFunction}
\title{Spatial correlation function as in 'ncf' package}
\usage{
spatialCorrelationFunction(mat, coords, resamp = 100, main = "", ...)
}
\arguments{
\item{mat}{A matrix.}

\item{coords}{.}

\item{resamp}{.}

\item{main}{.}

\item{\dots}{Additional graphical parameters.}
}
\value{
.
}
\description{
.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
