% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{SPI.36}
\alias{SPI.36}
\title{Sample data: Standardized Precipitation Index with 36 month accumulation period.}
\usage{
data(SPI.36)
}
\description{
A dataset containing 1200 rows (data length) and 252 columns.
}
\keyword{datasets}
