% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.diag.R
\name{taylor.diag}
\alias{taylor.diag}
\title{Taylor diagram}
\usage{
taylor.diag(
  ref,
  model,
  add = FALSE,
  col = "red",
  pch = 19,
  pos.cor = TRUE,
  xlab = "Standard deviation",
  ylab = "",
  main = "Taylor Diagram",
  show.gamma = TRUE,
  ngamma = 3,
  gamma.col = 8,
  sd.arcs = 0,
  cex.max = 1.5,
  sd.method = "sample",
  grad.corr.lines = c(0.2, 0.4, 0.6, 0.8, 0.9),
  pcex = 1,
  cex.axis = 1,
  normalize = TRUE,
  mar = c(4, 3, 4, 3),
  ...
)
}
\arguments{
\item{ref}{numeric vector - the reference values.}

\item{model}{numeric vector - the predicted model values.}

\item{add}{whether to draw the diagram or just add a point.}

\item{col}{the color for the points displayed.}

\item{pch}{the type of point to display.}

\item{pos.cor}{whether to display only positive (TRUE) or all values of correlation (FALSE).}

\item{xlab}{plot x axis labels.}

\item{ylab}{plot y axis labels.}

\item{main}{title for the plot.}

\item{show.gamma}{whether to display standard deviation arcs around the reference point (only for pos.cor=TRUE).}

\item{ngamma}{the number of gammas to display (default=3).}

\item{gamma.col}{color to use for the gamma arcs (only with pos.cor=TRUE).}

\item{sd.arcs}{whether to display arcs along the standard deviation axes (see Details).}

\item{cex.max}{scale of range of maximum sd.}

\item{sd.method}{whether to use the sample or estimated population SD.}

\item{grad.corr.lines}{the values for the radial lines for correlation values (see Details).}

\item{pcex}{character expansion/size for the plotted points.}

\item{cex.axis}{character expansion/size for the axis text.}

\item{normalize}{whether to normalize the models so that the reference has a standard deviation of 1.Default TRUE.}

\item{mar}{margins - only applies to the pos.cor=TRUE plot.}

\item{...}{Additional arguments passed to plot.}
}
\value{
The values of par that preceded the function. This allows the user to add points to the diagram, then restore the original values. This is only necessary when using the 0 to 1 correlation range.
}
\description{
Taylor diagram
}
\details{
The Taylor diagram is used to display the quality of model predictions against the reference values, typically direct observations.
A diagram is built by plotting one model against the reference, then adding alternative model points. If normalize=TRUE when plotting the first model, remember to set it to TRUE when plotting additional models.
Two displays are available. One displays the entire range of correlations from -1 to 1. Setting pos.cor to FALSE will produce this display. The -1 to 1 display includes a radial grid for the correlation values. When pos.cor is set to TRUE, only the range from 0 to 1 will be displayed. The gamma lines and the arc at the reference standard deviation are optional in this display.
Both the standard deviation arcs and the gamma lines are optional in the pos.cor=TRUE version. Setting sd.arcs or grad.corr.lines to zero or FALSE will cause them not to be displayed. If more than one value is passed for sd.arcs, the function will try to use the values passed, otherwise it will call pretty to calculate the values.
}
\examples{
# fake some reference data
ref<-rnorm(30,sd=2)
# add a little noise
model = model1<-ref+rnorm(30)/2
# add more noise
model2<-ref+rnorm(30)
par(pty="s")
# display the diagram with the better model
oldpar<-taylor.diag(ref,model1,cex.max = 1.2)
# now add the worse model
taylor.diag(ref,model2,add=TRUE,col="blue")
# get approximate legend position
lpos<-1.5*sd(ref)
# add a legend
legend(lpos,lpos,legend=c("Better","Worse"),pch=19,col=c("red","blue"))
# now restore par values
par(oldpar)
# show the "all correlation" display
taylor.diag(ref,model1,pos.cor=FALSE,cex.max = 1.2)
taylor.diag(ref,model2,add=TRUE,col="blue")
}
\references{
Taylor, K.E. (2001) Summarizing multiple aspects of model performance in a single diagram. Journal of Geophysical Research, 106: 7183-7192.
}
\author{
Olivier Eterradossi with modifications by Jim Lemon
}
