% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{waveletClustering}
\alias{waveletClustering}
\title{Perform wavelet clustering}
\usage{
waveletClustering(bwn)
}
\arguments{
\item{bwn}{A matrix of time series.}
}
\value{
A matrix of pairwise distances between wavelet spectra.
}
\description{
Compares wavelet spectra and generates an hierarchical tree for locations.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@pasteur.fr}).
}
