% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.6.scE2G.R
\name{sc_e2g_input}
\alias{sc_e2g_input}
\title{scE2G Pre-processing}
\usage{
sc_e2g_input(d_path, study_md)
}
\arguments{
\item{d_path}{Character string indicating the path
to a set of CellRanger files for data processing.}

\item{study_md}{A data frame containing sample
metadata variables specific to an individual study.}
}
\value{
Data frame containing a list of parameters
to use for scRNA-Seq data processing by Seurat.
}
\description{
Creates individual fragment and UMI files for processing
with scE2G pipeline, which is run separately using the CLI.
}
\examples{

sc_e2g_input("path/to/data/folder", df1)

}
