% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.6.scE2G.R
\name{sc_e2g_locusplot}
\alias{sc_e2g_locusplot}
\title{scE2G Locus Plot}
\usage{
sc_e2g_locusplot(
  path_proj,
  path_e2g,
  path_gtf,
  gene_name,
  genome = "hg38",
  col_ct = "CellType",
  track_window = 60000,
  track_promoter = 2000,
  thm_disc = col_univ(),
  thm_cont = col_grad(scm = 4),
  font_col = "grey25",
  font_size = 0.8,
  plot_int_anchors = FALSE,
  atac_interval = 500,
  atac_smooth = 3,
  font_panel_width = 1,
  show_scale_axis = FALSE
)
}
\arguments{
\item{path_proj}{Project directory location.}

\item{path_e2g}{Path to prediction results tsv gz file.}

\item{path_gtf}{Path to reference genome annotation file.}

\item{gene_name}{Gene name (character string).}

\item{genome}{Genome name.}

\item{col_ct}{Cell type column in prediction results file.}

\item{track_window}{Gene track plotting window, relative to
transcription start site for selected gene region.}

\item{track_promoter}{Promoter site window (in bp).}

\item{thm_disc}{Discrete color scheme to use.}

\item{thm_cont}{Continuous color scheme to use.}

\item{font_col}{Plot font color.}

\item{font_size}{Plot font size.}

\item{atac_interval}{bp interval for binning and smoothing
ATAC peaks.}

\item{atac_smooth}{Smoothing factor for ATAC peaks. Higher
numbers result in smoother peaks but may lose some information
for smaller neighboring peaks.}

\item{font_panel_width}{Width scaling factor for title panel.}

\item{show_scale_axis}{Show axis scales for ATAC peak signal
and heatmap interaction scores.}

\item{plot_int_achors}{Logical indicating whether to plot
anchors on the interaction track plot.}
}
\value{
Combined track plot visualizing enhancer-gene expression
interactions for a specified gene region.
}
\description{
Creates locus plot for selected gene region from an
scE2G results directory.
}
\examples{

sc_e2g_locusplot(
  path_proj = "/directory/to/results/",
  path_e2g = "path_proj/path/to/prediction.tsv.gz",
  path_gtf = "path_proj/path/to/gtffile.gtf.gz",
  gene_name = "gene name"
)

}
