% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.enrichment.R
\name{sc_enrichment}
\alias{sc_enrichment}
\title{scRNA-Seq Enrichment Analysis}
\usage{
sc_enrichment(l_deg, en_type, cstm_list, parl, core_perc)
}
\arguments{
\item{l_deg}{A dgea.output data frame created by sc.DGEA(). See sc.DGEA function documentation for details.}

\item{en_type}{Should enrichment statistics be calculated for GO terms or for a custom gene set list?
Type "GO" for GO enrichment or "cstm" to use a custom gene set list. If using a custom list, 
provide a text file containing gene names in column 1 and set IDs in column 2.}

\item{cstm_list}{Only used when conducting enrichment analysis based on a custom gene set list. 
Provide the path (relative to the working directory) to a custom gene set list for calculating enrichment statistics.}

\item{parl}{Logical indicating whether processing should be run in parallel (Linux and WSL2 only). Set to FALSE if running sequentially.}

\item{core_perc}{Percentage of available cores to use if running in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A list of Gene Ontology or custom gene set enrichment results per cell type for all comparisons
}
\description{
Performs enrichment analysis based on the Kolmogorov-Smirnov Test for a chosen list of DGEA results.
}
\examples{

# enr <- sc_enrichment(
# # Data frame containing DGEA results (created from sc.DGEA)
# l.deg = readRDS("analysis/object.dgea.result.rds")[[1]],
# # Enrichment type (Either "GO" or "cstm")
# en.type = "cstm",
# # If enrichment type is 'cstm', 
# # provide a custom gene set list (ignored when performing GO enrichment)
# cstm.list = setNames(
#   read.table(
#     "ref/gene.list.lipids.txt",header = T,sep = "\t"),
#     c("GENE","Set")
#     ),
# # run in parallel? (Set to FALSE if on Windows)
# parl = TRUE,
# # core percentage to use
# core.perc = 0.75
# )

}
