% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc_multiome_process}
\alias{sc_multiome_process}
\title{Process CellRanger Multiome Data Files}
\usage{
sc_multiome_process(
  df_par,
  i,
  p_name,
  m_cell,
  m_feat,
  atac_high,
  rna_high,
  rna_low,
  perc_mt,
  nsm_high,
  tss_low
)
}
\arguments{
\item{df_par}{Data frame of parameters to use for data processing.}

\item{i}{Numeric value for sample ID.}

\item{p_name}{Character string providing the project name.}

\item{m_cell}{Minimum cells per feature.}

\item{m_feat}{Minimum features per cell.}

\item{atac_high}{Upper limit for per cell ATAC counts.}

\item{rna_high}{Upper limit for per cell RNA counts.}

\item{rna_low}{Lower limit for per cell RNA counts.}

\item{perc_mt}{Upper limit for percentage of mitochondrial reads per cell.}

\item{nsm_high}{Upper limit for nucleosome signal. Used to filter
potential doublets, potential artifacts, etc.}

\item{tss_low}{Lower limit for TSS enrichment score. Used to filter
low quality cells.}
}
\value{
A list containing the processed multimodal Seurat object,
processing parameters, and QC plots.
}
\description{
Processes a single multiome sample for data integration.
}
\examples{

# proc_d_multiome <- sc_multiome_process(
#   # parameter list
#   l_params,
#   # sample ID
#   1,
#   # Project name
#   "Hiro.CF.multiome",
#   # Minimum cells per feature
#   5,
#   # Minimum features per cell
#   200,
#   # ATAC counts upper limit
#   100000,
#   # RNA counts upper limit
#   25000,
#   # RNA counts lower limit
#   1000,
#   # Percentage mitochondrial reads upper limit
#   20,
#   # Nucleosome signal upper limit
#   2.5,
#   # TSS enrichment lower limit
#   2
# )

}
