% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc_process_multiome_batch}
\alias{sc_process_multiome_batch}
\title{Batch Processing of CellRanger Multiome Files}
\usage{
sc_process_multiome_batch(
  df_par,
  p_name = "Project1",
  m_cell = 5,
  m_feat = 200,
  atac_high = 1e+05,
  rna_high = 25000,
  rna_low = 1000,
  perc_mt = 20,
  nsm_high = 2.5,
  tss_low = 2,
  parl = FALSE,
  core_num = NULL,
  pathmac = "/home/ncsteven/miniconda3/envs/r-env/bin/macs3"
)
}
\arguments{
\item{df_par}{Data frame of parameters to use for data processing.}

\item{p_name}{Character string providing the project name.}

\item{m_cell}{Minimum cells per feature.}

\item{m_feat}{Minimum features per cell.}

\item{atac_high}{Upper limit for per cell ATAC counts.}

\item{rna_high}{Upper limit for per cell RNA counts.}

\item{rna_low}{Lower limit for per cell RNA counts.}

\item{perc_mt}{Upper limit for percentage of mitochondrial reads per cell.}

\item{nsm_high}{Upper limit for nucleosome signal. Used to filter
potential doublets, potential artifacts, etc.}

\item{tss_low}{Lower limit for TSS enrichment score. Used to filter
low quality cells.}

\item{parl}{Logical indicating whether processing should be run in
parallel (Linux and WSL2 only).}

\item{core_num}{Number of available cores to use if running
in parallel (Linux and WSL2 only).}

\item{pathmac}{Path to MACS3 installation. This must be specified
to avoid errors in peak calling through Signac, which only supports
MACS2 by default. Specifying the path manually avoids this error.}
}
\value{
A processed list of 10X multiome sample files converted
into Seurat objects with a summary list of QC and processing details.
}
\description{
Processes a list of 10X multiome samples for downstream integration
using sc_multiome_process().
}
\examples{

# list_data <- sc_process_multiome_batch(df_par = scparam)

}
