% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.2.recluster.R
\name{sc_recluster}
\alias{sc_recluster}
\title{Reclustering Analysis}
\usage{
sc_recluster(
  so,
  rc_type = "GEX",
  ct_col = "CellType",
  slot1 = "RNA",
  slot2 = "ufy.peaks.corrected",
  red = "wnn.umap",
  sid = "Code",
  md_list,
  dim1 = 50,
  batch_cor = TRUE,
  cor_col,
  res1 = 0.5,
  h_w = 38,
  h_h = 18,
  fs_c = 6,
  fs_r = 9,
  cl_c = FALSE,
  cl_r = FALSE,
  rot_c = 45,
  col1 = col_grad(scm = 3),
  mdir = "analyze/",
  fsize = 45000
)
}
\arguments{
\item{so}{A Seurat object.}

\item{rc_type}{Data set type being reclustered (either "GEX" or "Mult").}

\item{ct_col}{Column containing cell type information.}

\item{slot1}{Assay from Seurat object subset to use (ex. "RNA").}

\item{slot2}{Secondary assay to use for reclustering
(only if rc_type is "Mult").}

\item{red}{Original dimension reduction to use for reclustering.}

\item{sid}{Sample ID column.}

\item{md_list}{A vector of character strings indicating metadata
columns for overlaying on a loadings plot.}

\item{dim1}{Number of dimensions to use in PCA.}

\item{batch_cor}{Should Harmony based batch correction be performed?
This parameter is automatically set to TRUE for multiome datasets.}

\item{cor_col}{Variables to account for in batch correction.}

\item{res1}{Clustering resolution to use for final cluster determination.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{col1}{Gradient color scheme to use
(must be exactly 4 colors in length).}

\item{mdir}{Directory for saving marker gene table.}

\item{fsize}{Future size for cell type predictions (in GB).}
}
\value{
A reclustered Seurat Object with summary plots, proportion tables,
and marker gene lists.
}
\description{
Performs reclustering and basic visualization
of a subsetted Seurat object subset.
Subsets are determined based on cell groups.
}
\examples{

# d_recluster <- sc_recluster(
#   so = d_gene,
#   md_list = c("Code", "Airway", "Batch"),
#   cor_col = c("Batch", "Code")
# )

}
