% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap}
\alias{sc_umap}
\title{Standard GEX/ATAC UMAP}
\usage{
sc_umap(
  so,
  md_var,
  slot1 = "umap",
  dims1 = "2D",
  col1 = col_univ(),
  pos_leg = "none",
  show_lab = TRUE
)
}
\arguments{
\item{so}{An input Seurat object.}

\item{md_var}{Name of the clustering column or overlay to plot.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}

\item{dims1}{Should the data be plotted in 2 or 3 dimensions?}

\item{col1}{Color scheme to use for plot overlay.}

\item{pos_leg}{Legend position (specify x and y coordinates or type "none")
(either "2D" or "3D").}

\item{show_lab}{Should labels be shown on the UMAP? Set to FALSE if plotting
intensity data.}
}
\value{
A UMAP with points grouped by a specific metadata column.
}
\description{
Generates a single UMAP with a specific
metadata overlay for a Seurat object
containing gene expression and/or chromatin accessibility data.
}
\examples{

# p_umap <- sc_umap(
#   so = d,
#   md_var = "CellType",
#   slot1 = "umap"
# )

}
