% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_umap_panel}
\alias{sc_umap_panel}
\title{UMAP panels}
\usage{
sc_umap_panel(so, md_list, slot1 = "umap", pos_leg = "none", show_lab = TRUE)
}
\arguments{
\item{so}{A Seurat object.}

\item{md_list}{A vector of metadata columns to plot.}

\item{slot1}{Name of the dimension reduction slot to plot.}

\item{pos_leg}{Legend position provided as a vector (default is "none").}

\item{show_lab}{Show labels on UMAP?}
}
\value{
A series of UMAPs with specified metadata overlays.
}
\description{
Generates a panel of UMAPs from a Seurat object.
}
\examples{

# p_umap <- sc_umap_panel(d_integrated,c("col1","col2","col3"),"wnn.umap")

}
