% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.volplot.R
\name{sc_volcano}
\alias{sc_volcano}
\title{Volcano Plot}
\usage{
sc_volcano(
  l_deg,
  ct,
  comp_name,
  gene_name = "GENE",
  filt_lab = NULL,
  diff_col = "log2FC",
  p_col = "H.qval",
  p_cut = 0.05,
  f_cut = 0.25,
  f_lim = 5,
  y_limit = 50,
  x_title
)
}
\arguments{
\item{l_deg}{A list of DGEA results returned by sc.DGEA().}

\item{ct}{A character string pattern for matching to specific cell types.}

\item{comp_name}{Comparison name, provided as a character string.}

\item{gene_name}{Column name containing gene or motif names.}

\item{filt_lab}{(optional) Select labels to include on plot.}

\item{diff_col}{Column name containing effect size.}

\item{p_col}{Column name containing adjust p-values.}

\item{p_cut}{Numeric value for p-value cutoff to indicate significance.}

\item{f_cut}{Fold change cutoff indicating high effect size.}

\item{f_lim}{Numeric value for fold change limits on the x-axis.}

\item{y_limit}{Upper-bound -log10 p-value to display on the y-axis.}

\item{x_title}{X-axis legend title, provided as a character string.}
}
\value{
A volcano plot for the chosen cell type and treatment comparison.
}
\description{
Generates a volcano plot from a DEG results object or data frame for
a chosen comparison and/or cell type.
}
\examples{

# p_vol <- sc_volcano(
#   l_deg = diff.output.activity,
#   ct = "1.Secretory",
#   comp_name = "KO vs. Control",
#   gene_name = "gene",
#   filt_lab = grepl("NKX|FOXA", diff.output.activity[["gene"]]),
#   diff_col = "avg_diff",
#   p_col = "p_val_adj",
#   p_cut = 0.005,
#   f_cut = 0.25,
#   f_lim = 6,
#   y_limit = 100,
#   x_title = "x-axis title"
# )

}
