% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_VT.R
\name{stepwise.VT}
\alias{stepwise.VT}
\title{Calculate stepwise high order VT in calibration}
\usage{
stepwise.VT(
  data,
  alpha = 0.1,
  mode = c("MRA", "MODWT", "AT"),
  wf,
  flag = c("biased", "unbiased"),
  detrend = F
)
}
\arguments{
\item{data}{A list of data, including response and predictors}

\item{alpha}{The significance level used to judge whether the sample estimate is significant. A default alpha value is 0.1.}

\item{mode}{A mode of variance transformation, i.e., MRA, MODWT, or AT}

\item{wf}{Wavelet family}

\item{flag}{Biased or Unbiased variance transformation}

\item{detrend}{Detrend the input time series or just center, default (F)}
}
\value{
A list of 2 elements: the column numbers of the meaningful predictors (cpy), and partial informational correlation (cpyPIC).
}
\description{
Calculate stepwise high order VT in calibration
}
\examples{
###Real-world example
mode <- switch(1,"MRA", "MODWT","a trous")
wf="d4"
station.id = 5 # station to investigate
SPI.12 <- SPEI::spi(rain.mon,scale=12)$fitted
lab.names <- colnames(obs.mon)
#plot.ts(SPI.12[,1:10])

x <- window(SPI.12[,station.id],start=c(1950,1),end=c(1979,12))
dp <- window(obs.mon[,lab.names],start=c(1950,1),end=c(1979,12))

data <- list(x=x,dp=matrix(dp, ncol=ncol(dp)))

dwt = stepwise.VT(data, mode=mode, wf=wf)

###plot transformed predictor before and after
par(mfrow=c(ncol(dp),1), mar=c(0,3,2,1))
for(i in 1:ncol(dp))
{
  ts.plot(cbind(dwt$dp[,i], dwt$dp.n[,i]), xlab="NA", col=1:2)
}
}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
