% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_VT.R
\name{stepwise.VT.val}
\alias{stepwise.VT.val}
\title{Calculate stepwise high order VT}
\usage{
stepwise.VT.val(data, J, dwt, mode)
}
\arguments{
\item{data}{A list of response and identifed predictors}

\item{J}{The decomposition level}

\item{dwt}{Output from dwt.vt(), including the transformation covariance}

\item{mode}{A mode of variance transfomration, i.e., MRA, MODWT, or AT}
}
\value{
A list of objects, including transformed predictors
}
\description{
Calculate stepwise high order VT
}
\examples{

}
