% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WASP-package.R
\docType{package}
\name{WASP-package}
\alias{WASP-package}
\title{WASP: WAvelet System Prediction}
\description{
The package WASP (variance transformation) is used for system modelling and prediction.
}
\details{
\tabular{ll}{
Package: \tab WASP\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2020-03-17\cr
License: \tab GPL-3\cr
}
}
\section{WASP functions}{

Variance transformation functions: dwt.vt, modwt.vt, at.vt and associated

K-nearest neighbor function: knn

Synthetic data generator functions: data.gen.SW,  data.gen.HL, data.gen.Rossler;
data.gen.ar1, data.gen.ar4, data.gen.ar9, data.gen.tar1, data.gen.tar2.
}

\references{
Jiang, Z., Sharma, A., & Johnson, F. (2020). Refining Predictor Spectral Representation Using Wavelet Theory for Improved Natural System Modeling. Water Resources Research, 56(3), e2019WR026962. doi:10.1029/2019wr026962

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge: Cambridge University Press.
}
\author{
Ze Jiang

Maintainer: Ze Jiang <ze.jiang@unsw.edu.au>
}
