% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WASP-package.R
\docType{package}
\name{WASP-package}
\alias{WASP-package}
\title{WASP: WAvelet System Prediction}
\description{
The package WASP (variance transformation) is used for system modelling and prediction.
}
\details{
\tabular{ll}{
Package: \tab WASP\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2020-03-17\cr
License: \tab GPL-3\cr
}
}
\section{WASP functions}{

Variance transformation functions: dwt.vt, modwt.vt, at.vt and associated

SPI calculation function

Synthetic data generator functions
}

\references{
Jiang, Z., Sharma, A., & Johnson, F. (2020). Refining predictor spectral representation using wavelet theory for improved natural system modelling. Water Resources Research, n/a(n/a), e2019WR026962. doi:10.1029/2019wr026962

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge: Cambridge University Press.
}
\author{
Ze Jiang

Maintainer: Ze Jiang zejiangunsw@gamil.com
}
